refactor: change structure of project files
