+++ nest-cli.json
@@ -1,7 +1,4 @@
 {
   "collection": "@nestjs/schematics",
-  "sourceRoot": "src",
-  "compilerOptions": {
-    "assets": ["/envs/*"]
-  }
+  "sourceRoot": "src"
 }
