+++ src/album/album.module.ts
@@ -2,7 +2,7 @@ import { forwardRef, Module } from "@nestjs/common";
 import { AlbumService } from './album.service';
 import { AlbumController } from './album.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { AlbumEntity } from './entity/album.entity';
+import { AlbumEntity } from './entities/album.entity';
 import { TrackModule } from "../track/track.module";
 import { FavoritesModule } from "../favorites/favorites.module";
