+++ src/app.controller.ts
@@ -1,5 +1,5 @@
 import { Controller } from '@nestjs/common';
-import { AppService } from 'src/app.service';
+import { AppService } from './app.service';
 
 @Controller()
 export class AppController {
