+++ src/artist/artist.module.ts
@@ -2,7 +2,7 @@ import { forwardRef, Module } from "@nestjs/common";
 import { ArtistService } from './artist.service';
 import { ArtistController } from './artist.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { ArtistEntity } from './entity/artist.entity';
+import { ArtistEntity } from './entities/artist.entity';
 import { TrackModule } from "../track/track.module";
 import { AlbumModule } from "../album/album.module";
 import { FavoritesModule } from "../favorites/favorites.module";
