+++ src/entries/artist.entity.ts
